Use catalog events and ticket types for checkout

+++ backend/src/main/java/com/paradise/event_ticket_system/checkout/CheckoutQuoteRequest.java
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.NotNull;
 import java.util.List;
 
 public record CheckoutQuoteRequest(
-        @NotNull Long eventId,
+        @NotNull Integer eventId,
         @NotEmpty List<@Valid TicketItemRequest> items,
         String discountCode
 ) {
