+++ backend/src/main/java/com/paradise/event_ticket_system/checkout/TicketItemRequest.java
@@ -4,7 +4,7 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 
 public record TicketItemRequest(
-        @NotNull Long ticketTypeId,
+        @NotNull Integer ticketTypeId,
         @NotNull @Min(1) Integer quantity
 ) {
 }
